dev-requirements.txt
mypy # Static type checker for Python
flake8 # Tool for style guide enforcement
types-mock # Type stubs for the mock library
pytest-cov # Plugin for pytest to measure code coverage